Add lambda-go to circleci
