+++ services/lambda-go/github.com/vleango/template.yml
@@ -7,7 +7,7 @@ Resources:
     Properties:
       Handler: bin/articles_index
       Runtime: go1.x
-      CodeUri: /go/.serverless/tech-writer.zip
+      CodeUri: /go/src/github.com/vleango/.serverless/tech-writer.zip
       MemorySize: 128
       Timeout: 3
       Events:
@@ -16,13 +16,6 @@ Resources:
           Properties:
             Path: /articles
             Method: get
-            RestApiId:
-              Ref: TechWriter
-        Event2:
-          Type: Api
-          Properties:
-            Path: /articles
-            Method: options
             RestApiId:
               Ref: TechWriter
   TechWriter:
@@ -222,7 +215,7 @@ Resources:
     Properties:
       Handler: bin/articles_create
       Runtime: go1.x
-      CodeUri: /go/.serverless/tech-writer.zip
+      CodeUri: /go/src/github.com/vleango/.serverless/tech-writer.zip
       MemorySize: 128
       Timeout: 3
       Events:
@@ -231,13 +224,6 @@ Resources:
           Properties:
             Path: /articles
             Method: post
-            RestApiId:
-              Ref: TechWriter
-        Event2:
-          Type: Api
-          Properties:
-            Path: /articles
-            Method: options
             RestApiId:
               Ref: TechWriter
   Articles_createLambdaPermission:
@@ -254,7 +240,7 @@ Resources:
     Properties:
       Handler: bin/articles_show
       Runtime: go1.x
-      CodeUri: /go/.serverless/tech-writer.zip
+      CodeUri: /go/src/github.com/vleango/.serverless/tech-writer.zip
       MemorySize: 128
       Timeout: 3
       Events:
@@ -279,7 +265,7 @@ Resources:
     Properties:
       Handler: bin/articles_update
       Runtime: go1.x
-      CodeUri: /go/.serverless/tech-writer.zip
+      CodeUri: /go/src/github.com/vleango/.serverless/tech-writer.zip
       MemorySize: 128
       Timeout: 3
       Events:
@@ -288,13 +274,6 @@ Resources:
           Properties:
             Path: '/articles/{id}'
             Method: post
-            RestApiId:
-              Ref: TechWriter
-        Event2:
-          Type: Api
-          Properties:
-            Path: /articles/{id}
-            Method: options
             RestApiId:
               Ref: TechWriter
   Articles_updateLambdaPermission:
@@ -311,7 +290,7 @@ Resources:
     Properties:
       Handler: bin/articles_destroy
       Runtime: go1.x
-      CodeUri: /go/.serverless/tech-writer.zip
+      CodeUri: /go/src/github.com/vleango/.serverless/tech-writer.zip
       MemorySize: 128
       Timeout: 3
       Events:
